+++ Storage/Storage/DTOs/WareHouseDto.cs
@@ -11,14 +11,14 @@ namespace Storage.DTOs
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public int TotalItem { get; set; }
+        public Int64 TotalItem { get; set; }
     }
 
     internal class WareHouse_DetailDto
     {
         public Guid WarehouseId { get; set; }
         public Guid Item_Id { get; set; }
-        public int Quantity { get; set; }
+        public Int64 Quantity { get; set; }
         public int Month { get; set; }
         public int Year { get; set; }
     }
